Refactor sims structure
