+++ .gitignore
@@ -3,5 +3,5 @@ __pycache__/
 *$py.class
 *.so
 simulation
-netrunner/target
+simlib/netrunner/target
 .idea/
