+++ simlib/mixnet-sims/src/log.rs
