+++ simlib/netrunner/src/network/behaviour.rs
