+++ simlib/netrunner/src/node/dummy_streaming.rs
