+++ simlib/netrunner/src/node/mod.rs
@@ -1,6 +1,5 @@
 #[cfg(test)]
 pub mod dummy_streaming;
-pub mod mix;
 
 // std
 use std::{
