+++ simlib/netrunner/src/streaming/io.rs
