+++ simlib/netrunner/src/streaming/naive.rs
