+++ simlib/netrunner/src/streaming/polars.rs
